+++ mobile/android/android-components/components/feature/syncedtabs/src/test/java/mozilla/components/feature/syncedtabs/controller/DefaultControllerTest.kt
@@ -1,9 +1,3 @@
-/*
- * This Source Code Form is subject to the terms of the Mozilla Public
- *  License, v. 2.0. If a copy of the MPL was not distributed with this
- *  file, You can obtain one at http://mozilla.org/MPL/2.0/.
- */
-
 /* This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
@@ -11,6 +5,7 @@
 package mozilla.components.feature.syncedtabs.controller
 
 import androidx.test.ext.junit.runners.AndroidJUnit4
+import kotlinx.coroutines.CompletableDeferred
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.test.TestCoroutineDispatcher
 import kotlinx.coroutines.test.runBlockingTest
@@ -23,6 +18,7 @@ import mozilla.components.feature.syncedtabs.storage.SyncedTabsStorage
 import mozilla.components.feature.syncedtabs.view.SyncedTabsView
 import mozilla.components.feature.syncedtabs.view.SyncedTabsView.ErrorType
 import mozilla.components.service.fxa.manager.FxaAccountManager
+import mozilla.components.service.fxa.sync.SyncReason
 import mozilla.components.support.test.mock
 import mozilla.components.support.test.rule.MainCoroutineRule
 import org.junit.After
@@ -64,7 +60,7 @@ class DefaultControllerTest {
             coroutineContext
         )
 
-        controller.syncTabs()
+        controller.refreshSyncedTabs()
 
         verify(view).startLoading()
         verify(view).stopLoading()
@@ -91,8 +87,51 @@ class DefaultControllerTest {
 
         `when`(storage.getSyncedTabs()).thenReturn(emptyList())
 
-        controller.syncTabs()
+        controller.refreshSyncedTabs()
 
         verify(view).onError(ErrorType.MULTIPLE_DEVICES_UNAVAILABLE)
+    }
+
+    @Test
+    fun `display synced tabs`() = runBlockingTest {
+        val controller = DefaultController(
+            storage,
+            accountManager,
+            view,
+            coroutineContext
+        )
+        val account: OAuthAccount = mock()
+        val constellation: DeviceConstellation = mock()
+        val state: ConstellationState = mock()
+
+        `when`(accountManager.authenticatedAccount()).thenReturn(account)
+        `when`(account.deviceConstellation()).thenReturn(constellation)
+        `when`(constellation.state()).thenReturn(state)
+        `when`(state.otherDevices).thenReturn(listOf(mock()))
+
+        `when`(storage.getSyncedTabs()).thenReturn(emptyList())
+
+        controller.refreshSyncedTabs()
+    }
+
+    @Test
+    fun `syncAccount refreshes devices and syncs`() = runBlockingTest {
+        val controller = DefaultController(
+            storage,
+            accountManager,
+            view,
+            coroutineContext
+        )
+        val account: OAuthAccount = mock()
+        val constellation: DeviceConstellation = mock()
+
+        `when`(accountManager.authenticatedAccount()).thenReturn(account)
+        `when`(account.deviceConstellation()).thenReturn(constellation)
+        `when`(constellation.refreshDevicesAsync()).thenReturn(CompletableDeferred(true))
+
+        controller.syncAccount()
+
+        verify(constellation).refreshDevicesAsync()
+        verify(accountManager).syncNowAsync(SyncReason.User, false)
     }
 }